+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Cover Mapping Profile Form with tests including field decorators (UIDATIMP-442)
 * Implement add new row functionality for the Field Mapping Profile details (UIDATIMP-497)
 * Change placement of repeatable decorator in the Inventory field mapping screens (UIDATIMP-518)
+* Implement delete row functionality for the Field Mapping Profile details (UIDATIMP-498)
 
 ### Bugs fixed:
 * Fix Item status list in field mapping profiles (UIDATIMP-515)